+++ src/directives/headerRowContainerTemplate/headerRowContainerTemplate.directive.ts
@@ -1,7 +1,8 @@
-import {Directive, TemplateRef, inject} from '@angular/core';
+import {Directive, Input, TemplateRef, inject} from '@angular/core';
 import {Func1} from '@jscrpt/common';
 
 import {GridRowContext} from '../../interfaces';
+import {rowColumnsAttribute} from '../../misc/utils';
 
 /**
  * Directive used for obtaining template for grid content renderer header row container
@@ -20,10 +21,19 @@ export class HeaderRowContainerTemplateSADirective
      */
     public template: TemplateRef<GridRowContext> = inject(TemplateRef<GridRowContext>);
 
+    //######################### public properties - inputs #########################
+
     /**
      * Predicate which controls rendering of row template, if not specified row is rendered
      */
+    @Input()
     public predicate: Func1<boolean, GridRowContext>|undefined|null;
+
+    /**
+     * Array of column ids to be rendered in this row, if not specified or null all available columns will be rendered
+     */
+    @Input({alias: 'headerRowContainerTemplate', transform: rowColumnsAttribute})
+    public columns: string[]|undefined|null;
 
     //######################### ng language server #########################